refactor instrumentExpress to be a lot less magic (#30)
I have taken out a bunch of the magic when instrumenting Express.

Instead of returning the required express I return an anonymous function (on which I copy all keys of the original express). When the method is called I initialize express and immediately register a middleware.

This should solve all issues no registering #28 or registering multiple times #9.