give event.js a better API (#22)
clean up the `event.js` API a bunch.  It now has a number of exported functions:

* `event.startRequest` creates a new toplevel event.  Creates a new `async_tracker`-tracked context, and initializes the `request_id` that is tracked by all sub-events.  calls `event.startEvent` to create the actual event.
* `event.finishRequest` sends an event using the payload for the toplevel event, and clears the tracked context.
* `event.startEvent`: marks the start time and source, and pushes the payload onto the current context's stack.
* `event.finishEvent`: adds the (user configurable column) event duration to the payload, optionally rolls up count + duration into the toplevel request event, and sends an event to libhoney.  Pops this event off the context's stack (along with any others on top of it in the stack - after warning the user) 

all instrumentations have been converted over to using this api.

`startEvent` / `finishEvent` replace the old pattern of manually noting the `startTime` and then calling a very complicated `sendEvent` api, doing the duration_ms calculation inline.

Now that I'm maintaining a stack of events, it's actually possible to warn the user of some instrumentation bug or (eventually) missing `finishEvent` call in user code.  Before there was no feedback we could offer - there just wouldn't be an event sent to honeycomb.

Currently the warning goes to the console which I don't like, but it's enough for now.  One additional feature I'll want to add (certainly before we document the API and mark it stable) is to record the call stack of the `startEvent` call, so we can give the user more info about which events didn't get sent.