add requestIdSource express config setting (#41)
Similar to the userContext config setting in #33, add another config setting for express to specify the request id to use (or how to find it.)

There are 3 ways to specify it:

* `express.requestIdSource = $string`
  looks for an http header `$string` and uses the value as the request id.

* `express.requestIdSource = $function-returning-string`
  calls that function and uses the return value as the request id.

* `express.requestIdSource = undefined` (or isn't defined)
  we look through headers for ones we know (currently `X-Request-ID` and `X-Amzn-Trace-Id`).

If we find a request from the above, we also add a field to the event named `meta.request_id_source` with either `$header http header` or `requestIdSource function` for the value.

We still generate request ids using `uuidv4` as an ultimate fallback if the above fails.  In that case `meta.request_id_source` isn't sent.