turn on manual tracing
We've already got containment of events that we're propagating in our context.  it's not a big leap to see that this enables in-process
tracing with pretty minimal changes.

There are some breaking schema changes, though.  no more "requestId".  it's "trace.trace_id", "trace.span_id", and "trace.parent_id" from
here on out.