propagate custom context to all events (#54)
Right now the customContext api adds to the current event. This is useless for everything but the request event, since in all other cases the event is sent before transfer is controlled back to user code.

Switch to keeping track of custom context separately from event payloads, and add the entirety of custom context to each event before we send it. For containing events (which request events are a simple 1-level case of), that means that we'll also add this context when popping the stack of events and sending.

By way of a terribly contrived example (also spelled out in docs/CustomContext.md):

```
1. request
  1.1 db call to look up user
  1.2 add custom context "user=some@email"
  1.3 SSR some react
```

the event generated by 1.1 won't include the custom context, since it's already been sent by the time we get to 1.2. 1.3 will contain the context, and so will 1 (since it contains 1.1 - 1.3, it will be sent after them.)

Con of this approach is that it exposes the order we send events. Pro is that it makes it possible to attach context to multiple events, and (while magic) it's at least explainable.

Fixes #35