Deterministic sampling (#56)
Add a deterministic sampler (sha1 request id, first 4 bytes treated as a BE uint32, compare that number to the sample rate's cutoff).

The sampling decision is made in `startRequest`.  If we decide to sample the request, we create the context and everything works as normal.  If we decide _not_ to sample it, `startRequest` returns null, and we don't create the context.  All the instrumentation treats this as a no-op, so for the cost of a single sampler invocation, we can sample the entire trace.