remove the explicit context first-parameter from event.startEvent (#63)
there's really no need for this parameter.  Everywhere that calls startEvent already makes sure we have a context (even if indirectly through `event.traceActive()` from #62 )

Remove it in favor of looking it up inside the body of startEvent.  removes the need for a local context variable (and even require-ing `async_tracker`) in a few files.