trace propagation (#59)
marshal trace context (trace_id, parent_id, and custom context) into an http header in both http/https instrumentations, and unmarshal it in the express instrumentation.

The format of the header value is:

```
  ${VERSION};${PAYLOAD}
```

with version 1 of the payload looking like:

```
  trace_id=${incoming trace id},parent_id=${incoming parent span id},context=${base64 encoded stringified JSON object}
```

Both `trace_id` and `parent_id` are required.  `context` can be missing without error, and any other `key=value` is ignored.  If `trace_id` or `parent_id` is missing, or if `context` fails to decode/parse as JSON, the context is dropped.