Add a public API (#66)
a lot going on here, but the api is in `lib/api/index.js`.  Could use some more comments there, but the idea is that it should cover a large % of external use-cases, as _well_ as be all an instrumentation author needs.

`Request` is renamed to `Trace`, `Event` is renamed to `Span`.  There are two new functions for those nouns, `with${Noun}`, which does both start/finish wrapped around a synchronous function.  The react-dom-server instrumentation now makes use of `withSpan`.

There's also high resolution timer support `startTimer`/`endTimer`/`withTimer`.

A number of functions from other modules are re-exported here with better names (in particular `bindFunctionToTrace` and `runWithoutTrace`).  These two are both used all over.

The bulk of the change is updating all instrumentations to use the new names (so many "`ev`"s got replaced by `trace` and `span`).

Fixes #12 
Fixes #40