[mongodb] deal with instrumented methods not being passed a callback (#76)
Turns out that mongodb collection methods that accept a callback have some _other_ return value if there's no callback passed.  Usually this is a promise, but for the `find*` methods, this can be a cursor.

The instrumentation was always creating a callback (to finish the span) and passing that to the wrapped function.  Needless to say, this causes things to break when the fluent api is used as it is in this little snippet from the mongodb code:

```javascript
var findOne = function(self, query, options, callback) {
  const cursor = self
    .find(query, options)
    .limit(-1)
    .batchSize(1);
...
```

That call to `self.find` calls our instrumented version without a callback.  since we passed the callback to `.find`, it didn't return the cursor which `findOne` depends on.

The fix is to make `wrap` a bit smarter.  I opted for a more dynamic approach - if there's a callback provided, we use the current behavior.  Otherwise we start the span as we normally would, and perform the original call.  If the return value has a `.then` method, we assume it's a promise and inject a then/catch handler that finishes the span and return that (it's still a promise usable by the calling code, and should be fine).  Otherwise we finish the span and return whatever the original function returns (this in particular covers the cursor case above in `findOne`).

We don't actually instrument cursor usage, so this means a `findOne` call will result in a single `find` span and a bit of a gap after it, as in:

![image](https://user-images.githubusercontent.com/24245/43491616-1a0348f0-94da-11e8-8184-f9f1468a129b.png)

that space between the ends of the `find` and `findOne` spans is the cursor stuff (I'm guessing the query doesn't actually get run until the cursor iterates to the first result?).