Support async/concurrent spans (#94)
Fixes #93 

Add `startAsyncSpan` to the API and only to the libhoney implementation for the time being (i.e. no tests).  It doesn't work the same as `startSpan`, unfortunately.  This is due to my understanding of how async context gets propagated by `async-hooks` - I may be wrong, so I'll keep digging there.

the signature is `startAsyncSpan(metadataContext, spanFn)`.  `metadataContext` is the same as with `startSpan`, and `spanFn` looks similar to `withSpan` but it's not.  it's called with a single argument - the asyncSpan itself.

From my motivating example in #93, here's `slowSpan` when using normal spans:

```
function slowSpan() {
  let asyncSpan = beeline.startSpan({ name: "slow" }); // should be startAsyncSpan here
  return new Promise(resolve => {
    setTimeout(() => {
      beeline.withSpan({ name: "slow-child" }, () => {});
      beeline.finishSpan(asyncSpan);
      resolve();
    }, 5000);
  });
}
```

and here's the converted form:

```
function slowSpan() {
  return beeline.startAsyncSpan(
    { name: "slow" },
    asyncSpan =>
      new Promise(resolve => {
        console.log("in slow promise");
        setTimeout(() => {
          beeline.withSpan({ name: "slow-child" }, () => {});
          beeline.finishSpan(asyncSpan);
          resolve();
        }, 5000);
      })
  );
}
```

The reason for passing the `asyncSpan` into this function is that the `finishSpan` call _must_ happen inside the async context created and installed around this inner function.  we create a new tracking context in `startAsyncSpan`, give it the same trace id, and install the async span as the root span in the new context's stack.  doing it this way means that other non-async child spans will be pushed onto the correct stacks and get the correct parent span id.