Add presend and sampler hooks implementation. (#125)
### Aim
To allow users of the library to define their own `samplerHook` function that takes an event's data and can make a sampling decision based on that. Also allow a `presendHook` function to be defined and called to enrich an event with further context.

### Reasoning (for samplerHook)
Dynamic sampling is something that some of the other Honeycomb libraries are capable of and it is really helpful!

### Discussion
Moved all sampling behaviour to `finishSpan()`, following the pattern demonstrated in the beelines-python library ([see here]). This will allow users to define custom sampler hooks that meet a standard interface. This change required refactor of existing deterministic sampler (and its tests) to fit with the new interface requirements.

### What's missing?
1. I haven't added any documentation. Usage docs don't appear to live in this repo.
2. There are no tests for what happens if you do not provide a samplerHook or a sampleRate. See roughly line 224 in `lib/api/libhoney.js` for more detail. I wasn't sure the best way to verify that `send()` was called rather than `sendPresampled()`.

[see here]: https://github.com/honeycombio/beeline-python/blob/master/beeline/__init__.py#L179