initial fastify instrumentation (#154)
Introduce `fastify` auto-instrumentation.  It registers two hooks before anything in user-code is run, one `onRequest` to start the trace, and one `onResponse` to send it.  Given just how pervasively async the api is and how things are registered, it's not as easy to ensure that context gets propagated correctly through hooks, so instead we store off the trace context in a Map and fetch it for every hook.  This means that there may be blips where context is lost, but each hook/handler should begin anew with the context.

Things this PR probably needs to include:

- [x] hooks
- [x] route handlers (`.get`, `.put`, etc)
- [ ] middleware - skipping this for now.  if we have requests, can add it via another PR.