PG instrumentation callbacks (#175)
Try to better handle the arguments being passed to the `query` function. It looks like having these three arguments have been pretty stable for many years. 

We need to decipher where in the arguments the callback lives. From reading the source, it looks like the options are these:

// query string, values, callback
// query string, callback
// Query like object
// Query like object, callback

Once we have deciphered where the correct callback lives we wrap it and pass it onto the wrapped `query` function.

This seems to work in my local manual testing. Looking into how to add some unit tests for this as well.

Resolves #173.