Make non-Honeycomb header trace parsing opt-in.
Before this change, we've been automatically parsing incoming Amazon
trace ID headers, and recently we started using the extracted values to
set trace.trace_id and trace.parent_id. This is potentially surprising
to users who are not using honeyaws to parse their ALB or ELB logs and
send the data to the same dataset, in that they'd get traces with
missing root spans.

This change requires that users of the node beeline specify the trace
source in the beeline initialization, e.g.:

```
require("honeycomb-beeline")({
    writeKey:    "foo",
    dataset:     "bar",
    serviceName: "foo-service",
    express: {
        traceIdSource: 'X-Amzn-Trace-Id',
    },
})
```
Note this is a breaking change so will have to be included in a major
release of the beeline.