Add null guard to addContext mock implementation
This PR brings the `addContext` implementation in `MockEventAPI` into
line with `LibhoneyEventAPI`.

The implementation of `addContext` in `LibhoneyEventAPI` checks that
a context exists before attempt to modify the context's stack.

The `MockEventAPI` implementation does not have this check. This can
cause tests to fail in services where the mock implementation is used
in test environments.