early return if the beeline has already been configured (#249)
Repeated calls to the toplevel configure function can yield errors like this:

```
The following modules were required before honeycomb-beeline: express
These modules will not be instrumented.  Please ensure honeycomb-beeline is required first.
```

This is because the toplevel configure function calls all the sub-configure functions, including our `instrumentation/` configure which checks for already-loaded modules, causing the warning above.

The function is meant to be idempotent, so early return if we're already configured.  We do, however, need a warning in this case as well - if the beeline configure call is passed arguments we should let the user know they aren't going to affect anything.

so, after this change:

```
const beeline = require("honeycomb-beeline");
beeline({ ... });
beeline({ ... }); // => "Beeline is already configured.  Further calls are allowed but should not be passed configuration options (as they will be ignored)."
```

warning feels kinda wordy, but /shrug