Handle pg pool clients (#272)
We found an issue when calling the promise returning version of query on a pg-pool client. These clients seem to behave a bit differently to both the query method on both pg.Pool and pg.Client. Both of those create a callback function before the the instrumentation is invoked, but in our case, the callback is created after instrumentation. The upshot is that no callback is passed to the query function instrumented by beeline, so even though the spans are started, they are never finished.

This PR adds a wrapper to promises returned by the query function to fix this issue. It also expands on the tests for the pg instrumentation by asserting that the results returned by the wrapped functions are correct. I've also added a bit to the README file to explain how to run the pg tests.

Co-authored-by: Mark Gaylard <mark.gaylard@rea-group.com>