rework of https.request, allows for WHATWG-URL obj
The honeycomb https.request wrapper instrumentation can't handle some of
the WHATWG-URL cases. For example, if you pass in a URL obj, options
(another object), and a callback, it will assume the first arg is the
options and the second is the callback. This will overwrite some of the
values (like the domain) and the call will fail.

For more info, see: https://nodejs.org/api/https.html#https_https_request_url_options_callback