Fix api.addContext so it works on the current span (not the root)
The documented behavior for `api.addContext` is that it adds fields to
the current span. However, the code was accessing index 0 of the context
stack, which will always be the *root* span. Tests previously did not
catch this because `api.addContext` was only exercised in a single-span
trace. A similar issue exists for `api.removeContext`, although it's
deprecated. But the patch is simple, so I applied it to both functions.

This probably wasn't caught sooner because `span.addContext` gets used
way more, particularly with asynchronous spans. As evidence, the
auto-instrumentation does not use `api.addContext` at all!

Note that running npm automatically bumped the package-lock.json, since
v2.7.0 of the Beeline had been released without changing the lock file.