Add trace-level rollup support (outreach fork)
Updates honeycomb-beeline with a new method, `incrementTraceRollup`,
which can be used to accumulate running counters over the course of a
request, to be reported on the top-level trace span when the request is
finished.

The original, unpatched honeycomb-beeline had some not-well-documented
suport for rollups, but it doesn't work with async spans, and most spans
are async.  This patched version will work with async spans, though it
does assume that even async spans don't outlive their parent trace.
That's generally true for us, but it's also a stronger assumption than
the upstream libraries are willing to make.  Our counts may be lower
than expected if that expectation is violated.